% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyByHapFreq.R
\name{classifyByHapFreq}
\alias{classifyByHapFreq}
\title{Classify By Haplotype Frequency}
\usage{
classifyByHapFreq(seq.df)
}
\arguments{
\item{seq.df}{data.frame of stratified and aligned sequences from \code{\link{gtypes2seqDF}}.}
}
\description{
Classify samples by haplotype frequency.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

