% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseToHaps.R
\name{collapseToHaps}
\alias{collapseToHaps}
\title{Collapse To Haplotypes}
\usage{
collapseToHaps(seq.df)
}
\arguments{
\item{seq.df}{data.frame of stratified and aligned sequences from \code{\link{gtypes2rfDF}}.}
}
\description{
Collapse sequence data.frame to unique haplotypes.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
