% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntRF.R
\name{confIntRF}
\alias{confIntRF}
\title{Confidence Intervals}
\usage{
confIntRF(rf, conf.level = 0.95)
}
\arguments{
\item{rf}{a \code{\link[randomForest]{randomForest}} object}

\item{conf.level}{confidence level for the \code{\link{binom.test}} confidence interval}
}
\description{
Calculate Random Forest confidence intervals
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

