% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gtype.rf.R
\name{gtype.rf}
\alias{gtype.rf}
\title{gtype Random Forest}
\usage{
gtype.rf(g, pairwise = FALSE, ...)
}
\arguments{
\item{g}{haploid \code{\link[strataG]{gtypes}} object with aligned sequences.}

\item{pairwise}{do analysis on all pairwise combinations of strata?}

\item{...}{arguments passed to \code{\link[randomForest]{randomForest}}.}
}
\value{
a list containing a data.frame of summary statistics (\code{smry}), and the
  \code{randomForest} object (\code{rf}). If \code{pairwise} is \code{TRUE} then
  the \code{rf} element is a list of \code{randomForest} results for each row in
  \code{smry}.
}
\description{
Conduct Random Forest on a gtypes object.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

