% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.haps.R
\name{make.haps}
\alias{make.haps}
\title{Make Haplotypes}
\usage{
make.haps(x)
}
\arguments{
\item{x}{data.frame of base pairs.}
}
\description{
Identify haplotypes in data.frame of aligned.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

