% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHaps.R
\name{makeHaps}
\alias{makeHaps}
\title{Make Haplotypes}
\usage{
makeHaps(x)
}
\arguments{
\item{x}{data.frame of base pairs.}
}
\description{
Identify haplotypes in data.frame of aligned.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

