% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctDiag.R
\name{pctDiag}
\alias{pctDiag}
\title{Extract Percent Diagnosable}
\usage{
pctDiag(rf, pd.vec = NULL)
}
\arguments{
\item{rf}{a \code{\link[randomForest]{randomForest}} object.}

\item{pd.vec}{a vector of of minimum vote values to return.}
}
\description{
Calculate Random Forest confidence intervals
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
