% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfSpeciesID.R
\name{rfSpeciesID}
\alias{rfSpeciesID}
\title{Random Forest Species ID}
\usage{
rfSpeciesID(g, ref.strata = NULL, unk.strata = NULL, ...)
}
\arguments{
\item{g}{haploid \code{\link[strataG]{gtypes}} object with aligned sequences.}

\item{ref.strata}{a character vector of 2 or more strata in \code{g} top use as reference samples.}

\item{unk.strata}{a character vector of strata representing unknown samples.}

\item{...}{arguments passed to \code{\link[randomForest]{randomForest}}.}
}
\value{
a list with the following elements:
\tabular{ll}{
  \code{pred} \tab vector of species predictions.\cr
  \code{prob} \tab matrix of species assignment probabilities.\cr
  \code{rf} \tab the \code{randomForest} model object.\cr
}
}
\description{
Predict species for unknown samples based on reference samples using Random Forest.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

